Create cia_1(dijikstrs)
this is the code for dijsktra.It can handle graphs consisting of cycles, but negative weights will cause this algorithm to produce incorrect results. Since Dijkstra follows a Greedy Approach, once a node is marked as visited it cannot be reconsidered even if there is another path with less cost or distance. This issue arises only if there exists a negative weight or edge in the graph.
So this algorithm fails to find the minimum distance in case of negative weights.